requirements.txt
geopandas>=0.12.0
pandas>=1.5.0
plotly>=5.0.0
matplotlib>=3.6.0
shapely>=2.0.0
fiona>=1.9.0
pyyaml>=6.0
streamlit>=1.20.0
scipy>=1.10.0
scikit-learn>=1.2.0